Refactor: btreePageSizeFromHeader encapsulates getting the page size from file header.
The idea is to make this available separately so that we can try all possible page
sizes when the database is encrypted. If the HMAC is correct and the page size
decoded from the header matches the currently attempted page size then it should be
safe to run with this value.