Disable out of memory fault simulation for now.
This is not usually a problem for our target platforms (Linux and Windows
with lots of RAM) and I'd rather fix the problems that are relevant for
us first.

This adds a new option --disable-oom-faults to test/tester.tcl to disable
OOM testing on invocation. You can still run the full test suite by not
passing this switch.